refactor: separating responsibilities for better readability and performance gains
- refactor: create PostgrestBuilder class
- refactor: update PostgrestTransformBuilder and PostgrestQueryBuilder select method to use cleanColumns from helper file

- feat: create PostgrestCount type
- feat: create helper file with cleanColumns method
