+++ src/index.ts
@@ -1,6 +1,6 @@
 import PostgrestClient from './PostgrestClient'
 import PostgrestFilterBuilder from './lib/PostgrestFilterBuilder'
 import PostgrestQueryBuilder from './lib/PostgrestQueryBuilder'
-import { PostgrestBuilder } from './lib/types'
+import { PostgrestBuilder } from './lib/PostgrestBuilder'
 
 export { PostgrestClient, PostgrestFilterBuilder, PostgrestQueryBuilder, PostgrestBuilder }
